api-gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
email-validator==2.1.0
requests==2.31.0
zeep==4.2.1
grpcio==1.62.2
grpcio-tools==1.62.2
protobuf==4.25.0
pika==1.3.2
pytest==8.3.4
pytest-mock==3.14.0
httpx==0.27.0